+++ CHANGELOG.md
@@ -1,11 +1,18 @@
-## 0.6.9 nightly
+## 0.7.0 nightly
+- Add pndm scheduler
+- Finish preview for discrete schedulers
+- Add new gif generator, save gif animations
+- Remove --model and automatically detect the model type before loading
+- Change gif to webp (lossy=100) for better quality
+- in progress...
+
+## 0.6.9
 - Fix client seed (reduce generator max seed from 19 to 16 digits)
 - Add config.json for user friendly config setup
 - Fix error message when mask exists but image is empty (return to txt2img)
 - Add latent preview (taesd)
 - Client moved to electron
 - Comfyui: visual improvements
-- in progress...
 
 ## 0.6.8
 - Back to the roots